requirements.txt
pandas==2.2.1
pydantic==2.6.4
PyMuPDF==1.24.1
pytest==8.1.1
pyyaml==6.0.1